jiji

+++ codes/referee_questions/check_correlation_variance/prepare_mock_cut/src/io.c
@@ -52,15 +52,6 @@ void load_pointing_list(int *N_plist, POINTING **plist){
 
 /* ----------------------------------------------------------------------- */
 
-// double normalize_PDF_Z(double z0, double z_min, double z_max){
-
-//     double pdf_norm;
-
-//     pdf_norm = 1.0 / ( 2.0 * z0 * ( tanh( z_max / (2.0 * z0) )
-//         - tanh( z_min / (2.0 * z0) ) ) );
-
-//     return pdf_norm;
-// }
 double integrate_Z(double z0, double z_min, double z_max){
 
     double integral;
@@ -73,16 +64,6 @@ double integrate_Z(double z0, double z_min, double z_max){
 }
 
 /* ----------------------------------------------------------------------- */
-
-// double normalize_PDF_R(double r0, double r_min, double r_max){
-
-//     double pdf_norm;
-
-//     pdf_norm = 1.0 / ( -r0 * ( exp(-r_max/r0)*(r_max + r0)
-//         - exp(-r_min/r0)*(r_min + r0) ) );
-
-//     return pdf_norm;
-// }
 
 double integrate_R(double r0, double r_min, double r_max){
 
@@ -158,13 +139,13 @@ void get_params( PARAMS *p, unsigned long int N ){
     density_const = (double)N / (thin_term + thick_term);
 
     /* get stars in thin disk */
-    double temp = density_const * thin_term;
-    p->N_thin = (int)temp;
+    long double temp = density_const * thin_term;
+    p->N_thin = (unsigned long int)temp;
 
     /* get stars in thick disk */
     /* add 1 to account for int roundoff */
     temp = density_const * thick_term;
-    p->N_thick = (int)temp + 1;
+    p->N_thick = (unsigned long int)temp + 1;
 
     fprintf(stderr, "%lu stars in the thin disk. \n", p->N_thin);
     fprintf(stderr, "%lu stars in the thick disk. \n", p->N_thick);
